extract_summary_functions

This is the main function for the fastGFPCA package. The function requires input data Y to be a dataframe in long format with variables id, index, and value to indicate subject IDs, observation times on the domain, and observations, respectively. The index must contain the same, equally spaced grid points for each subject. The number of functional principal components (FPCs) can either be specified directly (argument npc) or chosen based on the explained share of variance (pve). Families and link functions can be specified using the same syntax as the family argument from the lme4::glmer() function.

extract_summary_functions(
  mxFDAobject,
  r_vec = seq(0, 100, by = 10),
  extract_func = c(extract_univariate, extract_bivariate),
  summary_func = c(Kest, Lest, Gest),
  markvar,
  mark1,
  mark2 = NULL,
  edge_correction
)

Arguments

r_vec

Numeric vector of radii over which to evaluate spatial summary functions. Must begin at 0.

extract_func

Defaults to extract_univariate, which calculates univariate spatial summary functions. Choose extract_bivariate for bivariate spatial summary functions.

summary_func

Spatial summary function to calculate. Options are c(Kest, Lest, Gest) which denote Ripley's K, Besag's L, and nearest neighbor G function, respectively.

markvar

The name of the variable that denotes cell type(s) of interest. Character.

mark1

Character string that denotes first cell type of interest.

mark2

Character string that denotes second cell type of interest for calculating bivariate summary statistics. Not used when calculating univariate statistics.

edge_correction

Character string that denotes the edge correction method for spatial summary function. For Kest and Lest choose one of c("border", "isotropic", "Ripley", "translate", "none"). For Gest choose one of c("rs", "km", "han")

mxdata

Dataframe of cell-level multiplex imaging data. Should have variables x and y to denote x and y spatial locations of each cell.

analysis_vars

Optional list of variables to be retained for downstream analysis.

Value

A data.frame containing:

image_patient_id

the unique image id

r

the radius of values over which the spatial summary function is evaluated

sumfun

the values of the spatial summary function

csr

the values of the spatial summary function under complete spatial randomness

fundiff

sumfun - csr, positive values indicate clustering and negative values repulsion

References

Xiao, L., Ruppert, D., Zipunnikov, V., and Crainiceanu, C. (2016). Fast covariance estimation for high-dimensional functional data. Statistics and Computing, 26, 409-421. DOI: 10.1007/s11222-014-9485-x.

Author

Julia Wrobel julia.wrobel@cuanschutz.edu

Examples

# simulate data
set.seed(1001)