extract_univariate

Internal function called by extract_summary_functions to calculate a univariate spatial summary function for a single image.

Usage

extract_univariate(
  mximg,
  markvar,
  mark1,
  mark2,
  r_vec,
  func = c(Kest, Lest, Gest),
  edge_correction
)

Arguments

mximg

Dataframe of cell-level multiplex imaging data for a single image. Should have variables x and y to denote x and y spatial locations of each cell.

markvar

The name of the variable that denotes cell type(s) of interest. Character.

mark1

dummy filler, unused

mark2

dummy filler, unused

r_vec

Numeric vector of radii over which to evaluate spatial summary functions. Must begin at 0.

func

Spatial summary function to calculate. Options are c(Kest, Lest, Gest) which denote Ripley's K, Besag's L, and nearest neighbor G function, respectively.

edge_correction

Character string that denotes the edge correction method for spatial summary function. For Kest and Lest choose one of c("border", "isotropic", "Ripley", "translate", "none"). For Gest choose one of c("rs", "km", "han")

Value

A data.frame containing:

r

the radius of values over which the spatial summary function is evaluated

sumfun

the values of the spatial summary function

csr

the values of the spatial summary function under complete spatial randomness

fundiff

sumfun - csr, positive values indicate clustering and negative values repulsion

Author

Julia Wrobel julia.wrobel@emory.edu

Examples

# simulate data
set.seed(1001)