Multiplex imaging data from an ovarian cancer tumor microarray

This data is adapted from the VectraPolarisData Bioconductor package and comes from a tumor-microarray of tissue samples from 128 patients with ovarian cancer. There is one patient per subject.

ovarian_FDA

Format

ovarian_FDA

An mxFDA object with augmented ovarian cancer multiplex immunofluorescence data, and NN G(r) calculated:

Metadata

information about the spatial samples with column sample_key column in both

Spatial

cell-level information with x and y columns along with sample_key to link to Metadata

subject_key

column in Metadata that may have multiple sample_key values for each, akin to patient IDs

sample_key

column in both Metadata and Spatial that is a 1:1 with the samples (unique per sample)

univariate_summaries

univariate summary slot with nearest neighbor G calculared

bivariate_summaries

empty slot available for bivariate summaries

functional_pca

empty slot for functional PCA data of summaries

functional_cox

empty slot for functional models

Source

https://bioconductor.org/packages/release/data/experiment/html/VectraPolarisData.html

Details

Spatial summary functions of ovarian cancer multiplex imaging data.

This data is adapted from the VectraPolarisData Bioconductor package. Signal between the survival outcome and spatial summary functions has been augmented for teaching purposes. Spatial relationship is summarized using the nearest neighbor G function.