process_fcm

Fit a functional Cox regression model.

run_fcm(
  mxFDAobject,
  model_name,
  form,
  metric = "uni k",
  r = "r",
  value = "fundiff",
  afcm = FALSE,
  analysis_vars,
  quantile_transform = FALSE,
  smooth = FALSE,
  filter_cols = NULL,
  ...
)

Arguments

mxFDAobject

Dataframe of spatial summary functions from multiplex imaging data, in long format. Can be estimated using the function extract_summary_functions or provided separately.

model_name

character string to give the fit model in the functional cox slot

form

Formula to be fed to mgcv in the form of survival_time ~ x1 + x2. Does not contain functional predictor. Character valued. Data must contain censoring variable called "event".

metric

name of calculated spatial metric to use

r

Character string, the name of the variable that identifies the function domain (usually a radius for spatial summary functions). Default is "r".

value

Character string, the name of the variable that identifies the spatial summary function values. Default is "fundiff".

afcm

If TRUE, runs additive functional Cox model. If FALSE, runs linear functional cox model. Defaults to linear functional cox model.

analysis_vars

List of variables to be retained for downstream analysis, including variables from formula.

quantile_transform

Defaults to FALSE. If TRUE, a quantile transformation is applied to the functional predictor before modeling using the process_fcm function.

smooth

Option to smooth data using FPCA. Defaults to FALSE.

filter_cols

a named vector of factors to filter summary functions to in c(Derived_Column = "Level_to_Filter") format

...

Optional other arguments to be passed to fpca.face

knots

Number of knots for defining spline basis.

Value

A list which is a linear or additive functional Cox model fit. See mgcv::gam for more details.

Author

Julia Wrobel julia.wrobel@cuanschutz.edu

Alex Soupir alex.soupir@moffitt.org

Examples

# simulate data
set.seed(1001)